move straggler registry package
