+++ pkg/registry/core/controller/registry.go
