+++ pkg/registry/core/controller/strategy_test.go
